changed laneWidthBuilder to int

+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/LaneWidthBuilder.java
@@ -8,8 +8,8 @@ public class LaneWidthBuilder {
       throw new UnsupportedOperationException();
    }
 
-   public static BigDecimal laneWidth(BigDecimal jlw) {
-      return jlw.scaleByPowerOfTen(2);
+   public static int laneWidth(BigDecimal jlw) {
+      return (jlw.scaleByPowerOfTen(2).intValue());
    }
 
 }
